+++ packages/farcaster/package.json
@@ -13,6 +13,7 @@
     "@airstack/frames": "^1.2.0",
     "@farcaster/auth-client": "^0.3.0",
     "@farcaster/auth-kit": "^0.6.0",
+    "@packages/blockchain": "^0.0.0",
     "@packages/utils": "^1.0.0"
   }
 }
